+++ src/core/component/directives/safe-html/index.ts
@@ -22,18 +22,18 @@ import { ComponentEngine } from 'core/component/engines';
 export * from 'core/component/directives/safe-html/interface';
 
 ComponentEngine.directive('safe-html', (el: HTMLElement, {value, oldValue}: SafeHtmlDirectiveParams) => {
-	if (value == null || value === oldValue) {
+	if (value === oldValue) {
 		return;
 	}
 
 	let sanitized: string;
 
-	if (typeof value === 'string') {
-		sanitized = DOMPurify.sanitize(value, config.safeHtml);
+	if (Object.isPrimitive(value)) {
+		sanitized = DOMPurify.sanitize(formatValue(value), config.safeHtml);
 
 	} else {
 		sanitized = DOMPurify.sanitize(
-			value.value,
+			formatValue(value.value),
 
 			{
 				...config.safeHtml,
@@ -46,4 +46,12 @@ ComponentEngine.directive('safe-html', (el: HTMLElement, {value, oldValue}: Safe
 	}
 
 	el.innerHTML = sanitized;
+
+	/**
+	 * Formats the input value to a string for sanitization
+	 * @param value
+	 */
+	function formatValue(value: SafeHtmlDirectiveParams['value']): string {
+		return value == null ? '' : String(value);
+	}
 });