+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.??.? (2024-??-??)
+
+#### :rocket: New Feature
+
+* Supported the insertion of a primitive as a value `component/directives/safe-html`
+
 ## v3.69.1 (2024-04-09)
 
 #### :bug: Bug Fix